+++ flake.nix
@@ -6,59 +6,58 @@
       url = "github:m15a/flake-fennel-tools";
       inputs.nixpkgs.follows = "nixpkgs";
     };
-    fnldoc.url = "sourcehut:~m15a/fnldoc";
+    fnldoc = {
+      url = "sourcehut:~m15a/fnldoc";
+      inputs.nixpkgs.follows = "nixpkgs";
+    };
   };
 
   outputs = inputs @ { self, nixpkgs, flake-utils, fennel-tools, ... }:
-    flake-utils.lib.eachDefaultSystem (system:
+    ({
+      overlays.default = import ./nix/overlay.nix {
+        shortRev =
+          self.shortRev or
+          self.dirtyShortRev or
+          self.lastModified or
+          "unknown";
+      };
+    } // flake-utils.lib.eachDefaultSystem (system:
       let
         pkgs = import nixpkgs {
           inherit system;
           overlays = [
             fennel-tools.overlays.default
             inputs.fnldoc.overlays.default
+            self.overlays.default
+            (import ./nix/ci.nix)
           ];
         };
-
-        mkCICheckShell = { gnumake, fennel, faith, fnldoc }:
-          pkgs.mkShell {
-            buildInputs = [ gnumake fennel fennel.lua faith fnldoc ];
-            FENNEL_PATH = "${faith}/bin/?";
-          };
-
-        builder = { fennelVariant, luaVariant }:
-          let
-            fennelName =
-              if fennelVariant == "stable"
-              then "fennel-${luaVariant}"
-              else "fennel-${fennelVariant}-${luaVariant}";
-          in {
-            name = "ci-check-shell-${fennelName}";
-            value = pkgs.callPackage mkCICheckShell {
-              fennel = pkgs.${fennelName};
-              faith = pkgs.faith-unstable;
-            };
-          };
-
-        ci-check-shells = pkgs.lib.listToAttrs
-          (map builder (pkgs.lib.attrsets.cartesianProductOfSets {
-            fennelVariant = [ "stable" "unstable" ];
-            luaVariant = [ "lua5_1" "lua5_2" "lua5_3" "lua5_4" "luajit" ];
-          }));
       in
-      {
+      rec {
+        packages = rec {
+          inherit (pkgs) bumpfnl;
+          default = bumpfnl;
+        };
+
+        apps = with flake-utils.lib;
+          builtins.mapAttrs
+            (name: _: mkApp { drv = self.packages.${system}.${name}; })
+            packages;
+
+        checks = packages;
+
         devShells = {
-          inherit (ci-check-shells)
-          ci-check-shell-fennel-lua5_1
-          ci-check-shell-fennel-lua5_2
-          ci-check-shell-fennel-lua5_3
-          ci-check-shell-fennel-lua5_4
-          ci-check-shell-fennel-luajit
-          ci-check-shell-fennel-unstable-lua5_1
-          ci-check-shell-fennel-unstable-lua5_2
-          ci-check-shell-fennel-unstable-lua5_3
-          ci-check-shell-fennel-unstable-lua5_4
-          ci-check-shell-fennel-unstable-luajit;
+          inherit (pkgs)
+            ci-check-shell-fennel-lua5_1
+            ci-check-shell-fennel-lua5_2
+            ci-check-shell-fennel-lua5_3
+            ci-check-shell-fennel-lua5_4
+            ci-check-shell-fennel-luajit
+            ci-check-shell-fennel-unstable-lua5_1
+            ci-check-shell-fennel-unstable-lua5_2
+            ci-check-shell-fennel-unstable-lua5_3
+            ci-check-shell-fennel-unstable-lua5_4
+            ci-check-shell-fennel-unstable-luajit;
 
           default =
             let
@@ -67,18 +66,18 @@
             in
             pkgs.mkShell {
               buildInputs = [
-                pkgs.gnumake
                 fennel
                 fennel.lua
-                pkgs.fnlfmt-unstable
-                pkgs.fennel-ls-unstable
-                pkgs.faith-unstable
+                faith
                 pkgs.fnldoc
+                pkgs.fennel-ls-unstable
+                pkgs.fnlfmt-unstable
+                pkgs.gnumake
               ] ++ (with fennel.lua.pkgs; [
                 readline
               ]);
               FENNEL_PATH = "${faith}/bin/?";
             };
           };
-      });
+      }));
 }